+++ packages/terra-paginator/CHANGELOG.md
@@ -3,8 +3,14 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.18.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies
+
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 
 2.17.0  - (July 11, 2019)
 ------------------